reorganization

+++ include/utils/comb2vec.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstddef>
 #include <vector>
 
 /**
@@ -8,15 +9,15 @@
  * @return A vector of indices
  */
 template <typename T>
-std::vector<int> comb2vec(T comb) {
+std::vector<size_t> comb2vec(T comb) {
   static_assert(std::is_integral<T>::value, "T must be an integral type");
-  std::vector<int> vec;
+  std::vector<size_t> vec;
   while (comb) {
     T x = comb & -comb;  // get the rightmost bit
     if constexpr (sizeof(T) <= sizeof(int))
-      vec.push_back(__builtin_ctz(x));
+      vec.push_back(static_cast<size_t>(__builtin_ctz(x)));
     else if constexpr (sizeof(T) <= sizeof(long long))
-      vec.push_back(__builtin_ctzll(x));
+      vec.push_back(static_cast<size_t>(__builtin_ctzll(x)));
     else
       static_assert(sizeof(T) <= sizeof(int), "T is too large");
     comb &= comb - 1;  // clear the rightmost bit
